added Menger curvature test program
git-svn-id: https://femdk.l5.pk.edu.pl/bstools/BSTools/trunk@116 4ce36bd9-3a78-4973-a166-986cbda2e52d

+++ test/testg2/test6d.c
@@ -294,7 +294,7 @@ int main ()
     gh_DestroyDomaind ( domain );
   }
 
-  printf ( "Scratch memory used: %d bytes\n", pkv_MaxScratchTaken() );
+  printf ( "Scratch memory used: %d bytes\n", (int)pkv_MaxScratchTaken() );
   pkv_DestroyScratchMem ();
   exit ( 0 );
